Moved Url format methods onto ReturnDataManager
