+++ NavigationJS/src/StateController.ts
@@ -1,7 +1,7 @@
-﻿import ConverterFactory = require('./ConverterFactory');
-import Crumb = require('./Crumb');
+﻿import Crumb = require('./Crumb');
 import CrumbTrailManager = require('./CrumbTrailManager');
 import NavigationData = require('./NavigationData');
+import ReturnDataManager = require('./ReturnDataManager');
 import router = require('./router');
 import settings = require('./settings');
 import State = require('./config/State');
@@ -27,7 +27,7 @@ class StateController {
                 StateContext.previousDialog = StateContext.previousState.parent;
             CrumbTrailManager.returnData = {};
             if (data[settings.returnDataKey])
-                CrumbTrailManager.returnData = ConverterFactory.parseReturnData(data[settings.returnDataKey], StateContext.previousState);
+                CrumbTrailManager.returnData = ReturnDataManager.parseReturnData(data[settings.returnDataKey], StateContext.previousState);
             CrumbTrailManager.crumbTrail = data[settings.crumbTrailKey];
             StateContext.data = this.parseData(data, state);
             CrumbTrailManager.buildCrumbTrail();
@@ -127,7 +127,7 @@ class StateController {
         var newData = {};
         for (var key in data) {
             if (key !== settings.previousStateIdKey && key !== settings.returnDataKey && key !== settings.crumbTrailKey)
-                newData[key] = ConverterFactory.parseURLString(key, data[key], state);
+                newData[key] = ReturnDataManager.parseURLString(key, data[key], state);
         }
         NavigationData.setDefaults(newData, state.defaults);
         return newData;
